[test] 게임 매니저 연결하고 라이프 감소 및 UI 확인
게임 매니저 연결하고 라이프 감소 및 UI 확인

+++ Assets/Script/UI/Life/GameLifeUI.cs
@@ -14,7 +14,7 @@ public class GameLifeUI : MonoBehaviour, IGameLifeObserver
     public void OnGameLifeChanged(int currentLife, int maxLife)
     {
         _lifeText.text = currentLife.ToString() + "/" + maxLife.ToString();
-        _lifebar.fillAmount = currentLife / maxLife;
+        _lifebar.fillAmount = (float)currentLife / maxLife;
     }
 
     private void Start()
